skills/aurora-1412/arxiv-reader/requirements.txt
langchain>=1.2.9
langchain-openai>=1.1.7
requests>=2.31.0
python-dotenv>=1.0.0
arxiv>=2.1.0
arxiv-to-prompt

